refactor: separate sign up form schema from sign up form hook
Create `logInFormValidation.ts` file in `src/validation` folder,
And Separate sign up form schema (`signUpFormSchema`)
from `useSignUpFormLogic` hook into this file.
